requirements.txt
langchain==0.2.7
langchain_community==0.2.7
langchain_core==0.2.13
langchain_google_genai==1.0.7
langchain_openai==0.1.15
langchain_pinecone==0.1.1
langchain_text_splitters==0.2.2
pinecone==4.0.0
python-dotenv==1.0.1
streamlit==1.36.0
streamlit_chat==0.1.1
tiktoken==0.7.0
fastapi==0.111.0
nltk==3.8.1
jinja2==3.1.4
uvicorn==0.30.1
tqdm==4.66.4
langchainhub==0.1.20
pinecone-client[grpc]==4.1.2
openai==1.35.13
unstructured==0.14.10
